requirements.txt
absl-py==0.15.0
aiohttp==3.7.4.post0
appdirs==1.4.3
async-timeout==3.0.1
attrs==20.3.0
blessed==1.19.1
CacheControl==0.12.6
cachetools==4.2.4
captum==0.5.0
certifi==2019.11.28
chardet==3.0.4
click==8.0.3
cloudpickle==1.6.0
colorama==0.4.3
configparser==5.0.2
contextlib2==0.6.0
cycler==0.11.0
distlib==0.3.0
distro==1.4.0
docker-pycreds==0.4.0
eli5==0.11.0
fire==0.4.0
fonttools==4.34.4
fsspec==2021.10.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.0
google-auth-oauthlib==0.4.6
gpustat==1.0.0
graphviz==0.16
grpcio==1.41.0
html5lib==1.0.1
idna==2.8
imageio==2.22.2
ipaddr==2.2.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.4.4
lightning-bolts==0.5.0
llvmlite==0.35.0
lockfile==0.12.2
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.5.3
medmnist==2.1.0
msgpack==0.6.2
multidict==5.2.0
networkx==2.8.7
numba==0.52.0
numpy==1.20.1
nvidia-ml-py==11.495.46
oauthlib==3.1.1
packaging==20.3
pandas==1.2.3
pathtools==0.1.2
pep517==0.8.2
Pillow==8.1.2
progress==1.5
promise==2.3
protobuf==3.19.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
PyHessian==0.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pytorch-lightning==1.4.9
pytz==2021.1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
scikit-image==0.19.3
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.11.2
sentry-sdk==1.4.3
shap==0.39.0
shortuuid==1.0.1
six==1.14.0
slicer==0.0.7
smmap==5.0.0
subprocess32==3.5.4
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2022.10.10
torch==1.10.1+cu113
torchaudio==0.10.1+cu113
torchmetrics==0.5.1
torchvision==0.11.2+cu113
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.25.8
wandb==0.12.5
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
yarl==1.7.0
yaspin==2.1.0
